feat: unify Cover/Repaint/MusicEnhancer into single EnhancePanel (#776)
- Create EnhancePanel.tsx with Cover/Repaint tab toggle and smart mode inference
- Delete CoverModal.tsx, RepaintModal.tsx, MusicEnhancerPanel.tsx
- Unify uiStore state: enhancerTarget replaces coverClipId/repaintClipId/musicEnhancerOpen
- Update ClipBlock, AIToolsSubmenu, AppShell, Toolbar to use unified openEnhancer
- All 2232 tests passing, zero TypeScript errors

Closes #776
