+++ src/components/generation/EnhancePanel.tsx
@@ -1,0 +1,686 @@
+import { useState, useEffect, useCallback, useRef } from 'react';
+import { useProjectStore } from '../../store/projectStore';
+import { useGenerationStore } from '../../store/generationStore';
+import { useUIStore, getBottomPanelHeight } from '../../store/uiStore';
+import { generateCoverClip } from '../../services/generationPipeline';
+import { generateRepaintClip } from '../../services/generationPipeline';
+import { modelSupportsTaskType } from '../../services/aceStepApi';
+import { DualRangeSlider } from '../ui/DualRangeSlider';
+import { Z } from '../../utils/zIndex';
+import type { RepaintMode } from '../../types/api';
+
+const ENHANCER_BASE_BOTTOM = 60;
+
+type ConsistencyLevel = 'low' | 'medium' | 'high';
+const CONSISTENCY_VALUES: Record<ConsistencyLevel, number> = {
+  low: 0.25,
+  medium: 0.5,
+  high: 0.75,
+};
+
+function fmt(s: number) {
+  return `${s.toFixed(2)}s`;
+}
+
+interface SessionEntry {
+  id: string;
+  label: string;
+  timestamp: number;
+}
+
+interface ResultEntry {
+  id: string;
+  clipId: string;
+  title: string;
+  duration: string;
+  timestamp: number;
+}
+
+export function EnhancePanel() {
+  const enhancerOpen = useUIStore((s) => s.enhancerOpen);
+  const enhancerTarget = useUIStore((s) => s.enhancerTarget);
+  const closeEnhancer = useUIStore((s) => s.closeEnhancer);
+  const isGenerating = useGenerationStore((s) => s.isGenerating);
+  const getClipById = useProjectStore((s) => s.getClipById);
+  const project = useProjectStore((s) => s.project);
+  const bottomPanelHeight = useUIStore(getBottomPanelHeight);
+  const dynamicBottom = ENHANCER_BASE_BOTTOM + bottomPanelHeight;
+
+  const clip = enhancerTarget ? getClipById(enhancerTarget.clipId) : null;
+  const track = enhancerTarget
+    ? project?.tracks.find((t) => t.id === enhancerTarget.trackId) ?? null
+    : null;
+
+  // Local mode state — initialized from enhancerTarget.mode but user can toggle
+  const [mode, setMode] = useState<'cover' | 'repaint'>('cover');
+
+  // Cover fields
+  const [caption, setCaption] = useState('');
+  const [lyrics, setLyrics] = useState('');
+  const [consistency, setConsistency] = useState<ConsistencyLevel>('medium');
+  const [createNew, setCreateNew] = useState(true);
+
+  // Repaint fields
+  const [selStart, setSelStart] = useState(0);
+  const [selEnd, setSelEnd] = useState(0);
+  const [prompt, setPrompt] = useState('');
+  const [globalCaption, setGlobalCaption] = useState('');
+  const [repaintMode, setRepaintMode] = useState<RepaintMode>('balanced');
+  const [repaintStrength, setRepaintStrength] = useState(0.5);
+
+  // Sessions & results
+  const [sessions, setSessions] = useState<SessionEntry[]>([]);
+  const [results, setResults] = useState<ResultEntry[]>([]);
+  const [activeSessionId, setActiveSessionId] = useState<string | null>(null);
+  const sessionCounterRef = useRef(0);
+
+  // Initialize form when enhancerTarget changes
+  useEffect(() => {
+    if (enhancerTarget && clip) {
+      setMode(enhancerTarget.mode);
+
+      // Cover fields
+      setCaption(clip.prompt ?? '');
+      setLyrics(clip.lyrics ?? '');
+      setConsistency('medium');
+      setCreateNew(true);
+
+      // Repaint fields
+      const clipStart = clip.startTime;
+      const clipEnd = clip.startTime + clip.duration;
+      const rangeStart = enhancerTarget.range?.start ?? clipStart;
+      const rangeEnd = enhancerTarget.range?.end ?? clipEnd;
+      setSelStart(rangeStart);
+      setSelEnd(rangeEnd);
+      setPrompt(clip.prompt ?? '');
+      setGlobalCaption(clip.globalCaption ?? project?.globalCaption ?? '');
+      setRepaintMode('balanced');
+      setRepaintStrength(0.5);
+
+      // Create initial session
+      const sessionId = `session-${Date.now()}`;
+      sessionCounterRef.current = 1;
+      setSessions([{
+        id: sessionId,
+        label: 'Enhancement 1',
+        timestamp: Date.now(),
+      }]);
+      setActiveSessionId(sessionId);
+      setResults([]);
+    }
+  }, [enhancerTarget?.clipId]);
+
+  // Escape to close
+  useEffect(() => {
+    if (!enhancerOpen) return;
+    const handleEsc = (e: KeyboardEvent) => {
+      if (e.key === 'Escape') { e.stopPropagation(); closeEnhancer(); }
+    };
+    window.addEventListener('keydown', handleEsc);
+    return () => window.removeEventListener('keydown', handleEsc);
+  }, [enhancerOpen, closeEnhancer]);
+
+  const handleNewSession = useCallback(() => {
+    sessionCounterRef.current += 1;
+    const sessionId = `session-${Date.now()}`;
+    const entry: SessionEntry = {
+      id: sessionId,
+      label: `Enhancement ${sessionCounterRef.current}`,
+      timestamp: Date.now(),
+    };
+    setSessions((prev) => [entry, ...prev]);
+    setActiveSessionId(sessionId);
+    if (clip) {
+      setCaption(clip.prompt ?? '');
+      setLyrics(clip.lyrics ?? '');
+      setPrompt(clip.prompt ?? '');
+    }
+    setConsistency('medium');
+    setRepaintMode('balanced');
+    setRepaintStrength(0.5);
+    setResults([]);
+  }, [clip]);
+
+  const handleRangeChange = useCallback((s: number, e: number) => {
+    setSelStart(s);
+    setSelEnd(e);
+  }, []);
+
+  // Cover generation
+  const handleCoverGenerate = useCallback(async () => {
+    if (!enhancerTarget || isGenerating) return;
+    const coverStrength = CONSISTENCY_VALUES[consistency];
+    const resultId = `result-${Date.now()}`;
+    setResults((prev) => [...prev, {
+      id: resultId,
+      clipId: enhancerTarget.clipId,
+      title: caption || 'Untitled enhancement',
+      duration: '--:--',
+      timestamp: Date.now(),
+    }]);
+    await generateCoverClip({
+      clipId: enhancerTarget.clipId,
+      caption,
+      lyrics,
+      coverStrength,
+      createNew,
+    });
+  }, [enhancerTarget, caption, lyrics, consistency, createNew, isGenerating]);
+
+  // Repaint generation
+  const handleRepaintGenerate = useCallback(async () => {
+    if (!enhancerTarget || isGenerating) return;
+    const resultId = `result-${Date.now()}`;
+    setResults((prev) => [...prev, {
+      id: resultId,
+      clipId: enhancerTarget.clipId,
+      title: prompt || 'Untitled repaint',
+      duration: '--:--',
+      timestamp: Date.now(),
+    }]);
+    await generateRepaintClip({
+      clipId: enhancerTarget.clipId,
+      repaintStart: selStart,
+      repaintEnd: selEnd,
+      prompt,
+      globalCaption: globalCaption || undefined,
+      repaintMode,
+      repaintStrength,
+    });
+  }, [enhancerTarget, selStart, selEnd, prompt, globalCaption, repaintMode, repaintStrength, isGenerating]);
+
+  const handleGenerate = mode === 'cover' ? handleCoverGenerate : handleRepaintGenerate;
+
+  if (!enhancerOpen) return null;
+
+  // No-selection guidance screen
+  if (!enhancerTarget) {
+    return (
+      <div
+        data-testid="enhance-panel"
+        className="fixed left-1/2 -translate-x-1/2 w-[780px] bg-[#1e1e22] border border-[#3a3a3a] rounded-xl shadow-2xl text-xs text-zinc-200 p-8 text-center transition-[bottom] duration-200 ease-out"
+        style={{ zIndex: Z.panel, bottom: `${dynamicBottom}px` }}
+      >
+        <div className="flex items-center justify-between mb-6">
+          <span className="text-sm font-semibold text-white">Enhance</span>
+          <button
+            data-testid="enhance-close-btn"
+            onClick={closeEnhancer}
+            className="text-zinc-400 hover:text-zinc-200 transition-colors text-base leading-none"
+          >
+            ✕
+          </button>
+        </div>
+        <svg className="w-10 h-10 text-zinc-600 mx-auto mb-3" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.5">
+          <rect x="3" y="3" width="18" height="18" rx="2" strokeDasharray="4 2" />
+          <path d="M8 12h8M12 8v8" strokeLinecap="round" />
+        </svg>
+        <p className="text-zinc-400 text-[13px] mb-1">First, create a selection on the canvas</p>
+        <p className="text-zinc-600 text-[11px]">Use Cmd/Ctrl+drag on the timeline to select a region, or right-click a clip</p>
+        <div className="mt-6">
+          <button
+            onClick={closeEnhancer}
+            className="px-5 py-2 rounded-lg bg-[#2a2a2e] hover:bg-[#333338] text-zinc-300 text-[11px] font-medium transition-colors"
+          >
+            Cancel
+          </button>
+        </div>
+      </div>
+    );
+  }
+
+  const hasAudio = !!(clip?.isolatedAudioKey || clip?.cumulativeMixKey);
+  const coverSupported = modelSupportsTaskType('cover');
+  const repaintSupported = modelSupportsTaskType('repaint');
+  const modeSupported = mode === 'cover' ? coverSupported : repaintSupported;
+  const canGenerate = hasAudio && modeSupported && !isGenerating && !!(clip && track);
+
+  const clipStart = clip?.startTime ?? 0;
+  const clipEnd = (clip?.startTime ?? 0) + (clip?.duration ?? 0);
+  const totalDur = project?.totalDuration ?? clipEnd;
+
+  // Accent colors per mode
+  const accent = mode === 'cover' ? 'teal' : 'rose';
+  const accentBg = mode === 'cover' ? 'bg-teal-600' : 'bg-rose-600';
+  const accentBgHover = mode === 'cover' ? 'hover:bg-teal-500' : 'hover:bg-rose-500';
+  const accentBadgeBg = mode === 'cover' ? 'bg-teal-700/60' : 'bg-rose-700/60';
+  const accentBadgeText = mode === 'cover' ? 'text-teal-200' : 'text-rose-200';
+
+  return (
+    <div
+      data-testid="enhance-panel"
+      className="fixed left-1/2 -translate-x-1/2 w-[820px] max-h-[60vh] bg-[#1e1e22] border border-[#3a3a3a] rounded-xl shadow-2xl flex text-xs text-zinc-200 overflow-hidden transition-[bottom] duration-200 ease-out"
+      style={{ zIndex: Z.panel, bottom: `${dynamicBottom}px` }}
+    >
+      {/* Left Sidebar — Session History */}
+      <div data-testid="enhance-history" className="w-[150px] min-w-[150px] border-r border-[#3a3a3a] flex flex-col bg-[#1a1a1e]">
+        <div className="px-3 pt-3 pb-2">
+          <button
+            data-testid="enhance-new-session-btn"
+            onClick={handleNewSession}
+            className="w-full flex items-center justify-center gap-1.5 px-3 py-2 rounded-lg bg-[#2a2a2e] hover:bg-[#333338] text-zinc-300 text-[11px] font-medium transition-colors"
+          >
+            <span className="text-sm leading-none">+</span>
+            New Enhance
+          </button>
+        </div>
+        <div className="flex-1 overflow-y-auto px-1.5 pb-2 space-y-0.5">
+          {sessions.map((s) => (
+            <button
+              key={s.id}
+              onClick={() => setActiveSessionId(s.id)}
+              className={`w-full text-left px-2.5 py-2 rounded-md text-[11px] transition-colors truncate ${
+                s.id === activeSessionId
+                  ? 'bg-[#2a2a2e] text-zinc-100'
+                  : 'text-zinc-500 hover:bg-[#222226] hover:text-zinc-300'
+              }`}
+            >
+              {s.label}
+            </button>
+          ))}
+        </div>
+      </div>
+
+      {/* Center Panel — Controls */}
+      <div data-testid="enhance-controls" className="flex-1 min-w-0 flex flex-col border-r border-[#3a3a3a]">
+        {/* Header with mode toggle */}
+        <div className="flex items-center justify-between px-4 py-3 border-b border-[#3a3a3a]">
+          <div className="flex items-center gap-3">
+            <span className="text-sm font-semibold text-white">Enhance</span>
+            {/* Segmented tab toggle */}
+            <div className="flex bg-[#161618] rounded-md p-0.5" data-testid="enhance-mode-toggle">
+              <button
+                data-testid="enhance-mode-cover"
+                onClick={() => setMode('cover')}
+                className={`px-3 py-1 rounded text-[10px] font-bold uppercase tracking-wide transition-colors ${
+                  mode === 'cover'
+                    ? 'bg-teal-700/60 text-teal-200'
+                    : 'text-zinc-500 hover:text-zinc-300'
+                }`}
+              >
+                Cover
+              </button>
+              <button
+                data-testid="enhance-mode-repaint"
+                onClick={() => setMode('repaint')}
+                className={`px-3 py-1 rounded text-[10px] font-bold uppercase tracking-wide transition-colors ${
+                  mode === 'repaint'
+                    ? 'bg-rose-700/60 text-rose-200'
+                    : 'text-zinc-500 hover:text-zinc-300'
+                }`}
+              >
+                Repaint
+              </button>
+            </div>
+          </div>
+          <button
+            data-testid="enhance-close-btn"
+            onClick={closeEnhancer}
+            className="text-zinc-400 hover:text-zinc-200 transition-colors text-base leading-none"
+          >
+            ✕
+          </button>
+        </div>
+
+        <div className="flex-1 overflow-y-auto px-4 py-3 space-y-3 min-h-0">
+          {/* Source audio preview */}
+          <div className="bg-[#161618] rounded-lg px-3 py-3 border border-[#333]">
+            <p className="text-[10px] text-zinc-500 uppercase tracking-wide mb-1.5">Source</p>
+            {clip && track ? (
+              <>
+                <div className="flex items-center gap-2">
+                  <button
+                    data-testid="source-play-btn"
+                    className="w-7 h-7 flex items-center justify-center rounded-full bg-[#2a2a2e] text-zinc-400 hover:text-zinc-200 transition-colors"
+                    aria-label="Play source"
+                  >
+                    <svg className="w-3.5 h-3.5" viewBox="0 0 24 24" fill="currentColor"><path d="M8 5v14l11-7z" /></svg>
+                  </button>
+                  <div className="flex-1 min-w-0">
+                    <p className="text-[11px] font-medium text-zinc-200 truncate">
+                      {track.displayName ?? track.trackName}
+                    </p>
+                    <p className="text-[10px] text-zinc-500 truncate">{clip.prompt || '(no prompt)'}</p>
+                  </div>
+                </div>
+                {/* Waveform placeholder */}
+                <div className="mt-2 h-10 bg-[#1e1e22] rounded flex items-center justify-center">
+                  <div className="flex items-end gap-px h-6">
+                    {Array.from({ length: 40 }).map((_, i) => (
+                      <div
+                        key={i}
+                        className={`w-1 rounded-sm ${mode === 'cover' ? 'bg-teal-600/40' : 'bg-rose-600/40'}`}
+                        style={{ height: `${Math.max(2, Math.sin(i * 0.3) * 16 + Math.random() * 8)}px` }}
+                      />
+                    ))}
+                  </div>
+                </div>
+              </>
+            ) : (
+              <p className="text-[11px] text-zinc-500">No clip found</p>
+            )}
+            {clip && !hasAudio && (
+              <p className="text-[10px] text-amber-400 mt-2">
+                No audio generated yet — generate the clip first before enhancing.
+              </p>
+            )}
+            {!modeSupported && (
+              <p className="text-[10px] text-amber-400 mt-2">
+                The currently loaded model does not support {mode} generation.
+              </p>
+            )}
+          </div>
+
+          {/* === COVER MODE CONTROLS === */}
+          {mode === 'cover' && (
+            <>
+              {/* Lyrics */}
+              <div>
+                <label className="block text-[10px] font-medium text-zinc-500 uppercase tracking-wide mb-1">
+                  Lyrics
+                </label>
+                <textarea
+                  data-testid="enhance-lyrics-input"
+                  value={lyrics}
+                  onChange={(e) => setLyrics(e.target.value)}
+                  placeholder="Override lyrics for this enhancement..."
+                  rows={3}
+                  className="w-full bg-[#161618] border border-[#333] rounded-lg px-3 py-2 text-xs text-zinc-100 placeholder-zinc-600 resize-none focus:outline-none focus:border-teal-500/60 font-mono"
+                />
+              </div>
+
+              {/* Styles (caption) */}
+              <div>
+                <label className="block text-[10px] font-medium text-zinc-500 uppercase tracking-wide mb-1">
+                  Styles
+                </label>
+                <textarea
+                  data-testid="enhance-styles-input"
+                  value={caption}
+                  onChange={(e) => setCaption(e.target.value)}
+                  placeholder="e.g. jazz arrangement, acoustic guitar, slow tempo..."
+                  rows={2}
+                  className="w-full bg-[#161618] border border-[#333] rounded-lg px-3 py-2 text-xs text-zinc-100 placeholder-zinc-600 resize-none focus:outline-none focus:border-teal-500/60"
+                />
+              </div>
+
+              {/* Consistency */}
+              <div>
+                <label className="block text-[10px] font-medium text-zinc-500 uppercase tracking-wide mb-2">
+                  Consistency
+                </label>
+                <div className="flex gap-1" data-testid="enhance-consistency-toggle">
+                  {(['low', 'medium', 'high'] as ConsistencyLevel[]).map((level) => (
+                    <button
+                      key={level}
+                      onClick={() => setConsistency(level)}
+                      className={`flex-1 py-1.5 rounded-md text-[11px] font-medium capitalize transition-colors ${
+                        consistency === level
+                          ? 'bg-teal-600 text-white'
+                          : 'bg-[#161618] text-zinc-500 hover:bg-[#2a2a2e] hover:text-zinc-300'
+                      }`}
+                    >
+                      {level}
+                    </button>
+                  ))}
+                </div>
+              </div>
+
+              {/* Create new vs replace */}
+              <div className="flex items-center gap-3 pt-1">
+                <label className="flex items-center gap-2 cursor-pointer">
+                  <input
+                    type="checkbox"
+                    checked={createNew}
+                    onChange={(e) => setCreateNew(e.target.checked)}
+                    className="w-3.5 h-3.5 rounded border-[#444] bg-[#222] accent-teal-500"
+                  />
+                  <span className="text-[10px] text-zinc-400">Create new clip (leave original intact)</span>
+                </label>
+              </div>
+            </>
+          )}
+
+          {/* === REPAINT MODE CONTROLS === */}
+          {mode === 'repaint' && clip && (
+            <>
+              {/* Repaint range slider */}
+              <div className="bg-[#222]/60 rounded px-3 pt-2 pb-3 border border-[#3a3a3a]">
+                <div className="flex items-center justify-between mb-2">
+                  <span className="text-[10px] font-medium text-zinc-300">Repaint range</span>
+                  <span className="text-[10px] font-mono text-rose-300">
+                    {fmt(selStart)} — {fmt(selEnd)}
+                  </span>
+                </div>
+                <DualRangeSlider
+                  min={clipStart}
+                  max={clipEnd}
+                  startValue={selStart}
+                  endValue={selEnd}
+                  onChange={handleRangeChange}
+                  minSpan={0.1}
+                  step={0.01}
+                />
+                {/* Mini timeline diagram */}
+                <div className="relative mt-3" style={{ height: '28px' }}>
+                  <div
+                    className="absolute inset-x-0 bg-[#333] rounded"
+                    style={{ top: '12px', height: '4px' }}
+                  />
+                  {/* Clip region */}
+                  <div
+                    className="absolute bg-zinc-600/40 border border-zinc-500/40 rounded"
+                    style={{
+                      left: `${(clipStart / totalDur) * 100}%`,
+                      width: `${((clipEnd - clipStart) / totalDur) * 100}%`,
+                      top: '6px',
+                      height: '16px',
+                    }}
+                  />
+                  {/* Repaint region */}
+                  <div
+                    className="absolute bg-rose-600/50 border border-rose-500/70 rounded"
+                    style={{
+                      left: `${(selStart / totalDur) * 100}%`,
+                      width: `${((selEnd - selStart) / totalDur) * 100}%`,
+                      top: '6px',
+                      height: '16px',
+                    }}
+                  />
+                </div>
+                <div className="flex gap-4 mt-1">
+                  <span className="flex items-center gap-1 text-[8px] text-zinc-400">
+                    <span className="inline-block w-3 h-2 rounded-sm bg-zinc-600/50 border border-zinc-500/50" />
+                    Clip
+                  </span>
+                  <span className="flex items-center gap-1 text-[8px] text-rose-400">
+                    <span className="inline-block w-3 h-2 rounded-sm bg-rose-600/50 border border-rose-500/60" />
+                    Repaint region
+                  </span>
+                </div>
+              </div>
+
+              {/* Prompt */}
+              <div>
+                <label className="block text-[10px] font-medium text-zinc-400 uppercase tracking-wide mb-1">
+                  Prompt for this section
+                </label>
+                <textarea
+                  data-testid="enhance-repaint-prompt"
+                  value={prompt}
+                  onChange={(e) => setPrompt(e.target.value)}
+                  placeholder="Describe how this section should sound..."
+                  rows={3}
+                  className="w-full bg-[#222] border border-[#444] rounded px-2.5 py-2 text-xs text-zinc-100 placeholder-zinc-600 resize-none focus:outline-none focus:border-rose-500/60"
+                />
+              </div>
+
+              {/* Global caption */}
+              <div>
+                <label className="block text-[10px] font-medium text-zinc-400 uppercase tracking-wide mb-1">
+                  Global song description
+                  <span className="ml-1 normal-case font-normal text-zinc-600">(optional)</span>
+                </label>
+                <textarea
+                  data-testid="enhance-global-caption"
+                  value={globalCaption}
+                  onChange={(e) => setGlobalCaption(e.target.value)}
+                  placeholder="e.g. upbeat pop song..."
+                  rows={2}
+                  className="w-full bg-[#222] border border-[#444] rounded px-2.5 py-2 text-xs text-zinc-100 placeholder-zinc-600 resize-none focus:outline-none focus:border-rose-500/60"
+                />
+              </div>
+
+              {/* Repaint mode & strength */}
+              <div className="bg-[#222]/60 rounded px-3 py-2.5 border border-[#3a3a3a] space-y-2.5">
+                <div>
+                  <label className="block text-[10px] font-medium text-zinc-400 uppercase tracking-wide mb-1">
+                    Repaint mode
+                  </label>
+                  <div className="flex gap-1" data-testid="enhance-repaint-mode-toggle">
+                    {(['conservative', 'balanced', 'aggressive'] as const).map((rm) => (
+                      <button
+                        key={rm}
+                        onClick={() => setRepaintMode(rm)}
+                        className={`flex-1 px-2 py-1.5 rounded text-[10px] font-medium transition-colors ${
+                          repaintMode === rm
+                            ? 'bg-rose-600/80 text-white border border-rose-500'
+                            : 'bg-[#333] text-zinc-400 border border-[#444] hover:bg-[#3a3a3a]'
+                        }`}
+                      >
+                        {rm.charAt(0).toUpperCase() + rm.slice(1)}
+                      </button>
+                    ))}
+                  </div>
+                  <p className="text-[8px] text-zinc-600 mt-1">
+                    {repaintMode === 'conservative' && 'Maximum source preservation — subtle changes only.'}
+                    {repaintMode === 'balanced' && 'Tunable blend between source preservation and fresh generation.'}
+                    {repaintMode === 'aggressive' && 'Pure diffusion — fully regenerates the region.'}
+                  </p>
+                </div>
+
+                {repaintMode === 'balanced' && (
+                  <div>
+                    <div className="flex items-center justify-between mb-1">
+                      <label className="text-[10px] font-medium text-zinc-400">
+                        Repaint strength
+                      </label>
+                      <span className="text-[10px] font-mono text-rose-300">{repaintStrength.toFixed(2)}</span>
+                    </div>
+                    <input
+                      type="range"
+                      min={0}
+                      max={1}
+                      step={0.01}
+                      value={repaintStrength}
+                      onChange={(e) => setRepaintStrength(Number(e.target.value))}
+                      className="w-full h-1.5 accent-rose-500 cursor-pointer"
+                    />
+                    <div className="flex justify-between text-[8px] text-zinc-600 mt-0.5">
+                      <span>Preserve source</span>
+                      <span>Fresh generation</span>
+                    </div>
+                  </div>
+                )}
+              </div>
+
+              {mode === 'repaint' && (
+                <p className="text-[10px] text-zinc-600">
+                  Only the selected range will be regenerated. Audio outside the repaint region is preserved.
+                </p>
+              )}
+            </>
+          )}
+
+          {/* Enhance button */}
+          <button
+            data-testid="enhance-btn"
+            onClick={handleGenerate}
+            disabled={!canGenerate}
+            className={`w-full py-2.5 rounded-lg text-sm font-semibold transition-colors ${
+              canGenerate
+                ? `${accentBg} ${accentBgHover} text-white`
+                : 'bg-[#2a2a2e] text-zinc-500 cursor-not-allowed'
+            }`}
+          >
+            {isGenerating
+              ? (mode === 'cover' ? 'Enhancing...' : 'Repainting...')
+              : (mode === 'cover' ? 'Enhance' : 'Repaint Selection')
+            }
+          </button>
+        </div>
+      </div>
+
+      {/* Right Panel — Results */}
+      <div data-testid="enhance-results" className="w-[220px] min-w-[220px] flex flex-col bg-[#1a1a1e]">
+        <div className="px-3 py-3 border-b border-[#3a3a3a]">
+          <p className="text-[11px] font-semibold text-zinc-300 uppercase tracking-wide">Results</p>
+        </div>
+        <div className="flex-1 overflow-y-auto px-2 py-2 space-y-1">
+          {results.length === 0 ? (
+            <div className="flex flex-col items-center justify-center h-full text-center px-4">
+              <svg className="w-8 h-8 text-zinc-700 mb-2" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.5">
+                <path d="M9 18V5l12-2v13" strokeLinecap="round" strokeLinejoin="round" />
+                <circle cx="6" cy="18" r="3" />
+                <circle cx="18" cy="16" r="3" />
+              </svg>
+              <p className="text-[11px] text-zinc-600">
+                Enhanced results will appear here
+              </p>
+            </div>
+          ) : (
+            results.map((r) => (
+              <div
+                key={r.id}
+                className="flex items-center gap-2 px-2 py-2 rounded-md hover:bg-[#222226] transition-colors group"
+              >
+                <button
+                  className="w-6 h-6 flex items-center justify-center rounded-full bg-[#2a2a2e] text-zinc-400 hover:text-zinc-200 transition-colors flex-shrink-0"
+                  aria-label="Play result"
+                >
+                  <svg className="w-3 h-3" viewBox="0 0 24 24" fill="currentColor"><path d="M8 5v14l11-7z" /></svg>
+                </button>
+                <div className="flex-1 min-w-0">
+                  <p className="text-[11px] text-zinc-300 truncate">{r.title}</p>
+                  <p className="text-[10px] text-zinc-600">{r.duration}</p>
+                </div>
+                <button
+                  className="opacity-0 group-hover:opacity-100 text-zinc-500 hover:text-zinc-300 transition-opacity"
+                  aria-label="More options"
+                >
+                  <svg className="w-4 h-4" viewBox="0 0 24 24" fill="currentColor">
+                    <circle cx="12" cy="5" r="1.5" />
+                    <circle cx="12" cy="12" r="1.5" />
+                    <circle cx="12" cy="19" r="1.5" />
+                  </svg>
+                </button>
+              </div>
+            ))
+          )}
+        </div>
+
+        {/* Mini player */}
+        {results.length > 0 && (
+          <div className="border-t border-[#3a3a3a] px-3 py-2.5">
+            <div className="flex items-center gap-2">
+              <button className="text-zinc-500 hover:text-zinc-300 transition-colors" aria-label="Previous">
+                <svg className="w-3.5 h-3.5" viewBox="0 0 24 24" fill="currentColor"><path d="M6 6h2v12H6zm3.5 6l8.5 6V6z" /></svg>
+              </button>
+              <button className="text-zinc-300 hover:text-white transition-colors" aria-label="Play">
+                <svg className="w-4 h-4" viewBox="0 0 24 24" fill="currentColor"><path d="M8 5v14l11-7z" /></svg>
+              </button>
+              <button className="text-zinc-500 hover:text-zinc-300 transition-colors" aria-label="Next">
+                <svg className="w-3.5 h-3.5" viewBox="0 0 24 24" fill="currentColor"><path d="M16 18h2V6h-2zm-2-6L5.5 6v12z" /></svg>
+              </button>
+              <div className="flex-1 mx-1.5">
+                <div className="h-1 bg-[#2a2a2e] rounded-full">
+                  <div className="h-1 bg-teal-600 rounded-full w-0" />
+                </div>
+              </div>
+            </div>
+          </div>
+        )}
+      </div>
+    </div>
+  );
+}
