+++ tests/unit/clipContextMenuSplit.test.tsx
@@ -118,8 +118,7 @@ describe('ClipContextMenu AI Tools submenu', () => {
     renderMenu({
       clipAIContext: {
         onRegenerate: vi.fn(),
-        onCreateCover: vi.fn(),
-        onRepaint: vi.fn(),
+        onEnhance: vi.fn(),
         onAnalyze: vi.fn(),
         hasPrompt: true,
         isReady: true,
@@ -129,8 +128,7 @@ describe('ClipContextMenu AI Tools submenu', () => {
     fireEvent.mouseEnter(trigger);
     act(() => { vi.advanceTimersByTime(100); });
     expect(screen.getByText('Regenerate')).toBeTruthy();
-    expect(screen.getByText('Create Cover...')).toBeTruthy();
-    expect(screen.getByText('Repaint Selection...')).toBeTruthy();
+    expect(screen.getByText('Enhance...')).toBeTruthy();
     expect(screen.getByText('Analyze Audio...')).toBeTruthy();
   });
 });
